Refactor Utils file
