+++ app/backend/blueprints/backend_blueprint/chat.py
@@ -8,7 +8,8 @@ from error import error_response
 from models.chat import TextChatRequest
 from models.request_type import RequestType
 from quart import Blueprint, current_app, request
-from utils.utils import Utils
+from utils.request_handler import RequestHandler
+from utils.response_handler import ResponseHandler
 
 chat = Blueprint("chat", __name__, url_prefix="/chat")
 
@@ -18,16 +19,16 @@ async def chat_stream_endpoint():
     try:
         # Receive data from the client
         data = await request.get_json()
-        textChatRequest = TextChatRequest(**Utils.form_query_request(data).model_dump())
+        textChatRequest = TextChatRequest(**RequestHandler.form_query_request(data).model_dump())
         approach = cast(Approach, current_app.config[CONFIG_CHAT_APPROACH])
 
         result = await approach.run_stream(
-            messages=Utils.form_message(textChatRequest.chat_history, textChatRequest.query),
+            messages=RequestHandler.form_message(textChatRequest.chat_history, textChatRequest.query),
             profile=textChatRequest.profile,
             language=textChatRequest.language,
         )
 
-        response = await Utils.construct_streaming_response(result, RequestType.CHAT)
+        response = await ResponseHandler.construct_streaming_response(result, RequestType.CHAT)
         return response, 200
     except Exception as error:
         logging.error("Exception in /chat/stream. ", error)
@@ -39,16 +40,16 @@ async def chat_stream_endpoint():
 async def chat_endpoint():
     try:
         data = await request.get_json()
-        textChatRequest = TextChatRequest(**Utils.form_query_request(data).model_dump())
+        textChatRequest = TextChatRequest(**RequestHandler.form_query_request(data).model_dump())
         approach = cast(Approach, current_app.config[CONFIG_CHAT_APPROACH])
 
         result = await approach.run(
-            messages=Utils.form_message(textChatRequest.chat_history, textChatRequest.query),
+            messages=RequestHandler.form_message(textChatRequest.chat_history, textChatRequest.query),
             profile=textChatRequest.profile,
             language=textChatRequest.language,
         )
 
-        response = Utils.construct_non_streaming_response(result)
+        response = ResponseHandler.construct_non_streaming_response(result)
 
         return response, 200
     except Exception as error:
