+++ app/backend/blueprints/backend_blueprint/voice.py
@@ -7,7 +7,8 @@ from error import error_response
 from models.request_type import RequestType
 from models.voice import VoiceChatRequest
 from quart import Blueprint, current_app, request
-from utils.utils import Utils
+from utils.request_handler import RequestHandler
+from utils.response_handler import ResponseHandler
 
 voice = Blueprint("voice", __name__, url_prefix="/voice")
 
@@ -19,17 +20,17 @@ async def voice_endpoint():
         # Receive data from the client
         data = await request.get_json()
         # Extract data from the JSON message
-        voiceChatRequest = VoiceChatRequest(**Utils.form_query_request(data).model_dump())
+        voiceChatRequest = VoiceChatRequest(**RequestHandler.form_query_request(data).model_dump())
 
         # Send transcribed text and data to LLM
         approach = cast(Approach, current_app.config[CONFIG_CHAT_APPROACH])
         result = await approach.run_stream(
-            messages=Utils.form_message(voiceChatRequest.chat_history, voiceChatRequest.query),
+            messages=RequestHandler.form_message(voiceChatRequest.chat_history, voiceChatRequest.query),
             profile=voiceChatRequest.profile,
             language=voiceChatRequest.language.lower(),
         )
 
-        response = await Utils.construct_streaming_response(
+        response = await ResponseHandler.construct_streaming_response(
             result, RequestType.VOICE, voiceChatRequest.language.lower()
         )
         return response, 200
