+++ app/backend/utils/json_encoder.py
@@ -1,5 +1,9 @@
 import dataclasses
 import json
+import logging
+from typing import AsyncGenerator
+
+from error import error_dict
 
 
 class JSONEncoder(json.JSONEncoder):
@@ -12,3 +16,12 @@ class JSONEncoder(json.JSONEncoder):
         if dataclasses.is_dataclass(o) and not isinstance(o, type):
             return dataclasses.asdict(o)
         return super().default(o)
+
+    @staticmethod
+    async def format_as_ndjson(r: AsyncGenerator[dict, None]) -> AsyncGenerator[str, None]:
+        try:
+            async for event in r:
+                yield json.dumps(event, ensure_ascii=False, cls=JSONEncoder) + "\n"
+        except Exception as error:
+            logging.exception("Exception while generating response stream: %s", error)
+            yield json.dumps(error_dict(error))
